Get the release build working

+++ Generals/Code/GameEngine/Source/Common/System/GameMemory.cpp
@@ -3338,7 +3338,7 @@ void* operator new(size_t size, const char * fname, int)
 #ifdef MEMORYPOOL_DEBUG
 	return TheDynamicMemoryAllocator->allocateBytesImplementation((int)size, fname);
 #else
-	return TheDynamicMemoryAllocator->allocateBytesImplementation(size);
+	return TheDynamicMemoryAllocator->allocateBytesImplementation((int)size);
 #endif
 }
 
@@ -3366,7 +3366,7 @@ void* operator new[](size_t size, const char * fname, int)
 #ifdef MEMORYPOOL_DEBUG
 	return TheDynamicMemoryAllocator->allocateBytesImplementation((int)size, fname);
 #else
-	return TheDynamicMemoryAllocator->allocateBytesImplementation(size);
+	return TheDynamicMemoryAllocator->allocateBytesImplementation((int)size);
 #endif
 }
